requirements.txt
PyYAML==3.11
tai5-uan5-gian5-gi2-kang1-ku7==0.5.6
